Fix Mobile Center page duration tracking
Environment.TickCount can be negative, it wraps every time it reaches max value and is currently negative on our test devices...
Plus that part of the if was redundant with the boolean value returned by TryGetValue which already checks if the dictionary was actually containing a mapping for the hash.  Thus negative value or 0 are legit values here.